refactor: split AgentFactory into partials by responsibility
653 lines, past this repo's 400-line cap. Same breach #251 fixed on its
sibling in the same folder, one file over, and the same approach: a pure move,
so there is no behavioural risk to weigh at review time.

Three files, cut where the concerns actually part:

  AgentFactory.cs (231)            construction dependencies and every public
                                   entry point that does not start from skill
                                   ids - provider availability, creation from a
                                   built context, persistent provisioning,
                                   routed chat-client resolution
  AgentFactory.ChatClient.cs (288) given a resolved provider: the two
                                   construction paths, resilient-client
                                   substitution, and the middleware pipeline
                                   they share
  AgentFactory.FromSkills.cs (194) the entry points that start from skill ids,
                                   prerequisite validation, batch discovery

The trap the issue names is avoided: only the primary partial carries a
class-level <summary>, because the compiler merges class-level docs across
partials into one entry and publishes whichever compiles first, silently. The
other two say so in a comment.

Pure move proven three ways rather than asserted: 322 non-comment code lines
identical in both directions; all 31 member signatures identical including
modifiers and parameter defaults; and no partial adds a using relative to the
original, so no moved call site can have rebound.

Closes #273
